+++ lib/rorvswild/installer.rb
@@ -48,7 +48,9 @@ production:
   #   - Redis
   #   - Resque
   #   - Sidekiq
+  #
   # logger: log/rorvswild.log # By default it uses Rails.logger or Logger.new(STDOUT)
+  #
   # # Deployment tracking is working without any actions from your part if the Rails app
   # # is inside a Git repositoriy, is deployed via Capistrano.
   # # In the other cases, you can provide the following details.
@@ -57,6 +59,16 @@ production:
   #   description: <%= "Eventually if you have a description such as a Git message" %>
   #   author: <%= "Author's name of the deployment" %>
   #   email: <%= "emailOf@theAuthor.example" %>
+  #
+  # Sampling allows to send a fraction of jobs and requests.
+  # If your app is sending hundred of millions of requests per month,
+  # you will probably get the same precision if you send only a fraction of it.
+  # Thus, it decreases the bill at the end of the month. It's also a mitigation if
+  # your app is a target of a DoS. There are 2 parameters to dissociate requests and jobs.
+  # Indeed, for an app handling a lot of request but very few jobs, it makes sens to sample
+  # the former but not the latter.
+  # request_sampling_rate: 0.25 # 25% of requests are sent
+  # job_sampling_rate: 0.5 # 50% of jobs are sent
 YAML
     end
   end